rename notify -> notifications
Notify is already taken, see https://lib.rs/crates/notify.
